Updated JavaScript code to support the defaultSource prop
