+++ src/interface.js
@@ -1,0 +1,13 @@
+import { Image } from 'react-native'
+import PropTypes from 'prop-types'
+
+module.exports = {
+  name: 'FirebaseImage',
+  propTypes: {
+    ...Image.propTypes,
+    path: PropTypes.string,
+    timestamp: PropTypes.number,
+    resizeMode: PropTypes.oneOf(['cover', 'contain', 'stretch']),
+    defaultSource: PropTypes.number
+  }
+}
